Add S3/Valkey/Kafka benchmarks and make the stack addressable by hostname
Adds benchmark/, a set of scripts that measure throughput and latency
percentiles for the three data services against any host running the
stack:

    make bench HOST=nas.local

Shared measurement lives in benchmark/harness.py and benchmark/stats.py
so every phase is timed identically. Notable choices:

* Explicit worker threads, not a ThreadPoolExecutor. A pool only spawns a
  thread when none is idle, so with fast operations one thread drains
  every slice and --concurrency silently means 1.
* Throughput is ops over wall-clock elapsed, never over summed latency,
  which at concurrency 8 would report an eighth of the real rate.
* Warmup runs on the worker threads through the same code path, untimed,
  so the p99 is not simply the first request.
* Cleanup deletes the exact names the run created, from memory, never a
  prefix scan, so it cannot touch anything the benchmark did not write.

Pointing a client at a hostname exposed that Kafka was not reachable by
one. It binds 0.0.0.0 like everything else, but advertised
PLAINTEXT_HOST://localhost:9094, so a client on another machine
bootstrapped, was told to reconnect to localhost, and dialled its own
loopback until it timed out.

* KAFKA_ADVERTISED_HOST now sets the address the broker hands out,
  defaulting to localhost so single-host use is unchanged.
* The two listeners swap ports: clients get the conventional 9092 and
  the compose-internal listener moves to kafka:29092. Only 9092 is
  published; the internal port could never work from the host, since it
  advertises a name only containers resolve.
* The Kafka benchmark checks the advertised address before doing any
  work and prints the fix instead of hanging.

Audited the other services for the same class of problem: FLOCI_HOSTNAME
and Grafana's GF_SERVER_ROOT_URL are now templated the same way, both
defaulting to today's values.

.env.example ships BIND_ADDR=0.0.0.0, matching the compose fallback the
two previously disagreed on. A fresh copy is LAN-reachable, so the
credential warning next to it and in README's security section now leads
with changing the passwords.

No new dependencies: boto3, valkey and confluent-kafka were already
present, and the statistics are stdlib. benchmark/ sits outside
--cov=dashboard, so the 100% gate is unaffected; the pure helpers get 73
unit tests that need no containers.

+++ dashboard/clients.py
@@ -22,7 +22,7 @@ if TYPE_CHECKING:
 DEFAULT_TIMEOUT = 5
 
 
-def make_valkey_client(cfg: ValkeyConfig, timeout: int = DEFAULT_TIMEOUT) -> Any:
+def make_valkey_client(cfg: ValkeyConfig, timeout: float = DEFAULT_TIMEOUT) -> Any:
     import valkey
 
     return valkey.Valkey(host=cfg.host, port=cfg.port, socket_timeout=timeout)
